small bug fix
