+++ tedeous/utils.py
@@ -1,10 +1,13 @@
 """this one contain some stuff for computing different auxiliary things."""
 
-from typing import Tuple, List
+from typing import Tuple, List, Union
 from torch.nn import Module
-from SALib import ProblemSpec
+import datetime
+import os
+import shutil
 import numpy as np
 import torch
+from tedeous.device import check_device
 
 def create_random_fn(eps: float) -> callable:
     """ Create random tensors to add some variance to torch neural network.
@@ -86,6 +89,183 @@ def bcs_reshape(
                                         for i in range(bval_diff.shape[-1])])
 
     return bcs
+
+def remove_all_files(folder: str) -> None:
+    """ Remove all files from folder.
+
+    Args:
+        folder (str): folder name.
+    """
+    for filename in os.listdir(folder):
+        file_path = os.path.join(folder, filename)
+        try:
+            if os.path.isfile(file_path) or os.path.islink(file_path):
+                os.unlink(file_path)
+            elif os.path.isdir(file_path):
+                shutil.rmtree(file_path)
+        except Exception as e:
+            print('Failed to delete %s. Reason: %s' % (file_path, e))
+
+
+class CacheUtils:
+    """ Mixin class with auxiliary methods
+    """
+    def __init__(self):
+        try:
+            file = __file__
+        except:
+            file = os.getcwd()
+
+        self._cache_dir = os.path.normpath((os.path.join(os.path.dirname(file), '..', 'cache')))
+
+    def get_cache_dir(self):
+        """Get cache dir.
+
+        Returns:
+            str: cache folder directory.
+        """
+        return self._cache_dir
+
+    def set_cache_dir(self, string: str) -> None:
+        """ Change the directory of cache.
+
+        Args:
+            string (str): new cache directory.
+        """
+        self._cache_dir = string
+
+    def clear_cache_dir(self, directory: Union[str, None] = None) -> None:
+        """ Clear cache directory.
+
+        Args:
+            directory (str, optional): custom cache directory. Defaults to None.
+        """
+        if directory is None:
+            remove_all_files(self.cache_dir)
+        else:
+            remove_all_files(directory)
+
+    cache_dir = property(get_cache_dir, set_cache_dir, clear_cache_dir)
+
+    @staticmethod
+    def grid_model_mat(model: torch.Tensor,
+                       grid: torch.Tensor,
+                       cache_model: torch.nn.Module=None) -> Tuple[torch.Tensor, torch.nn.Module]:
+        """ Create grid and model for *NN or autograd* modes from grid
+            and model of *mat* mode.
+
+        Args:
+            model (torch.Tensor): model from *mat* method.
+            grid (torch.Tensor): grid from *mat* method.
+            cache_model (torch.nn.Module, optional): neural network that will 
+                                                     approximate *mat* model. Defaults to None.
+
+        Returns:
+            nn_grid (torch.Tensor): grid satisfying neural network inputs.
+            cache_model (torch.nn.Module): model satisfying the *NN, autograd* methods.
+        """
+        nn_grid = torch.vstack([grid[i].reshape(-1) for i in \
+                                range(grid.shape[0])]).T.float()
+        input_model = grid.shape[0]
+        output_model = model.shape[0]
+
+        if cache_model is None:
+            cache_model = torch.nn.Sequential(
+                torch.nn.Linear(input_model, 100),
+                torch.nn.Tanh(),
+                torch.nn.Linear(100, 100),
+                torch.nn.Tanh(),
+                torch.nn.Linear(100, 100),
+                torch.nn.Tanh(),
+                torch.nn.Linear(100, output_model)
+            )
+
+        return nn_grid, cache_model
+
+    @staticmethod
+    def mat_op_coeff(operator: dict) -> dict:
+        """ Preparation of coefficients in the operator of the *mat* method
+            to suit methods *NN, autograd*.
+
+        Args:
+            operator (dict): operator (equation dict).
+
+        Returns:
+            operator (dict): operator (equation dict) with suitable coefficients.
+        """
+        if not isinstance(operator, list):
+            operator = [operator]
+        for op in operator:
+            for label in list(op.keys()):
+                term = op[label]
+                if isinstance(term['coeff'], torch.Tensor):
+                    term['coeff'] = term['coeff'].reshape(-1, 1)
+                elif callable(term['coeff']):
+                    print("Warning: coefficient is callable,\
+                                    it may lead to wrong cache item choice")
+        return operator
+
+    def save_model(
+        self,
+        model: torch.nn.Module,
+        name: Union[str, None] = None) -> None:
+        """
+        Saved model in a cache (uses for 'NN' and 'autograd' methods).
+        Args:
+            model (torch.nn.Module): model to save.
+            (uses only with mixed precision and device=cuda). Defaults to None.
+            name (str, optional): name for a model. Defaults to None.
+        """
+
+        if name is None:
+            name = str(datetime.datetime.now().timestamp())
+        if not os.path.isdir(self.cache_dir):
+            os.mkdir(self.cache_dir)
+        parameters_dict = {'model': model.to('cpu'),
+                           'model_state_dict': model.state_dict()}
+
+        try:
+            torch.save(parameters_dict, self.cache_dir + '\\' + name + '.tar')
+            print('model is saved in cache')
+        except RuntimeError:
+            torch.save(parameters_dict, self.cache_dir + '\\' + name + '.tar',
+                       _use_new_zipfile_serialization=False)  # cyrrilic in path
+            print('model is saved in cache')
+        except:
+            print('Cannot save model in cache')
+
+    def save_model_mat(self, model: torch.Tensor,
+                       grid: torch.Tensor,
+                       cache_model: Union[torch.nn.Module, None] = None,
+                       name: Union[str, None] = None) -> None:
+        """ Saved model in a cache (uses for 'mat' method).
+
+        Args:
+            model (torch.Tensor): *mat* model
+            grid (torch.Tensor): grid from *mat* mode
+            cache_model (Union[torch.nn.Module, None], optional): model to save. Defaults to None.
+            name (Union[str, None], optional): name for a model. Defaults to None.
+        """
+
+        nn_grid, cache_model = self.grid_model_mat(model, grid, cache_model)
+        optimizer = torch.optim.Adam(cache_model.parameters(), lr=0.001)
+        model_res = model.reshape(-1, model.shape[0])
+
+        def closure():
+            optimizer.zero_grad()
+            loss = torch.mean((cache_model(check_device(nn_grid)) - model_res) ** 2)
+            loss.backward()
+            return loss
+
+        loss = np.inf
+        t = 0
+        while loss > 1e-5 and t < 1e5:
+            loss = optimizer.step(closure)
+            t += 1
+            print('Interpolate from trained model t={}, loss={}'.format(
+                    t, loss))
+
+        self.save_model(cache_model, name=name)
 
 
 class PadTransform(Module):
